requirements.txt
async-timeout==4.0.2
black==22.6.0
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
Deprecated==1.2.13
docutils==0.19
idna==3.3
Kivy==2.1.0
Kivy-examples==2.1.0
Kivy-Garden==0.1.5
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
Pillow==9.2.0
platformdirs==2.5.2
Pygments==2.12.0
pyparsing==3.0.9
python-dotenv==0.20.0
redis==4.3.4
requests==2.28.1
six==1.16.0
spotipy==2.20.0
tomli==2.0.1
typing_extensions==4.3.0
urllib3==1.26.11
wrapt==1.14.1
